FeliCa support, a rollup of years of changes
This patch originally years ago with taking the changes published at
https://lantaukwcounter.blogspot.com/2015/11/reading-octopus-card-ids-felica-with.html

and working them into the adafruit_pn532 library.

Sadly, I did not keep my patches clear and time has faded the memory of
what was actually done when and why.

The adafruit library has moved on (it now has fancier bus interface parts)
but there is still hope that this might be mergable one day.